requirements.txt
# A股量化回测框架依赖包

# 核心数据处理
pandas>=2.0.0
numpy>=1.24.0

# 数据可视化
matplotlib>=3.7.0
seaborn>=0.12.0

# 配置文件处理
pyyaml>=6.0

# 测试框架
pytest>=7.4.0
pytest-cov>=4.1.0

# 代码质量
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# 类型提示
typing-extensions>=4.7.0

# 日期时间处理
python-dateutil>=2.8.0

# 可选：数据获取
akshare>=1.11.0
